+++ Modules/Project.AllPlatforms.Utility/Public/Show-SDDL.ps1
@@ -46,7 +46,7 @@ TODO: additional work on function to make it more universal, see if we can make 
 #>
 function Show-SDDL
 {
-	[OutputType([string])]
+	[OutputType([string[]])]
 	[CmdletBinding(
 		HelpURI = "https://github.com/metablaster/WindowsFirewallRuleset/blob/develop/Modules/Project.AllPlatforms.Utility/Help/en-US/Show-SDDL.md"	)]
 	param (